docs(Input): move to components folder
