+++ src/components/Input/docs/InputExtendedExample.st.css
@@ -1,5 +1,5 @@
 :import {
-  -st-from: "../../src/components/Input/Input.st.css";
+  -st-from: "../Input.st.css";
   -st-default: TPAInput;
 }
