+++ src/components/Input/docs/InputExtendedExample.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import {Input, InputProps} from '../../src/components/Input';
+import {Input, InputProps} from '../Input';
 import extendedStyles from './InputExtendedExample.st.css';
 
 export const InputExtendedExample: React.FunctionComponent<InputProps> = (props) => <Input {...props} {...extendedStyles('root', {}, props)}/>;
